Ignore `IsAvailable` property in Bed and Room entities
The `BedConfiguration` class was updated to ignore the `IsAvailable`
property of the `Bed` entity, preventing it from being mapped to the
database. Similarly, the `RoomConfiguration` class was modified to
ignore the `IsAvailable` property of the `Room` entity. These changes
indicate that the `IsAvailable` property will now be handled outside
the database context, possibly through dynamic calculation or other
means.